Prefer adduser to useradd
The adduser command is supplied on Ubuntu, RedHat, and other popular distros and is usually the
recommended way to create new user accounts.  The useradd command is considered lower level and
typically doesn't perform all of the "usual" setup for new accounts.  The only reason we were
using it is because CoreOS doesn't supply the adduser command.